fix(distribution): address review feedback on threshold rank derivation
- Make prefix extraction infallible instead of fallible: destructure the
  first 8 bytes of the fixed-size hash with an array pattern, so there is
  no slice conversion, no allocation concern and no panic path at all.
  Returning a Result was considered and rejected: rank computation over a
  32-byte hash cannot fail, and a fallible signature would poison every
  caller for an unreachable error. Padding shorter inputs was also
  rejected: zero-padding biases short inputs toward high ranks, silently
  turning a misuse into a wrong tree shape. Taking the fixed-size array
  makes the invalid input unrepresentable instead.
- Take the branch factor as u64 (including the Distribution trait const)
  so the threshold math does not convert per call.
- Widen Rank to u64 to match the threshold arithmetic.
- Drop the explicit max_rank_for cap: integer division drives the
  threshold to zero on its own, so the loop terminates naturally at
  floor(log_m(2^64)) + 1. This also removes the m=1 non-termination
  hazard from the const fn; a debug_assert now documents that branch
  factors below 2 are out of domain.
- Seed the statistical tests with a fixed StdRng so they are
  deterministic instead of racing sigma tolerances against thread_rng.
- Add tests: exact threshold boundary goldens (including equality, which
  must not promote), a per-level conditional promotion probability test
  (the regression test for the bit-batch straddling bug this PR fixes),
  and a determinism test proving only the 8-byte prefix participates.
- Port the same tests to dialog-search-tree, which previously had none.

+++ rust/dialog-prolly-tree/src/distribution.rs
@@ -5,7 +5,7 @@ pub use geometric::*;
 use crate::KeyType;
 
 /// A rank determines how a tree's segments should be chunked
-pub type Rank = u32;
+pub type Rank = u64;
 
 /// A trait that may be implemented by any type that defines how to derive the
 /// [`Rank`] of a value in a tree
@@ -15,7 +15,7 @@ where
     Hash: HashType,
 {
     /// The branch factor used by this distribution strategy
-    const BRANCH_FACTOR: u32;
+    const BRANCH_FACTOR: u64;
 
     /// Compute the [`Rank`] of a value given its key
     // TODO: support tree state e.g., fn rank(state: &TreeState, key: &Key) -> Rank;
